requirements.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.78
botocore==1.31.78
certifi==2023.7.22
charset-normalizer==3.3.2
distro==1.8.0
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
jmespath==1.0.1
multidict==6.0.4
openai==1.2.2
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
pypdf==3.17.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.7.0
safetensors==0.4.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
telegram==0.0.1
tiktoken==0.5.1
tokenizers==0.14.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
youtube-transcript-api==0.6.1
